refactor(web): Unify ui.Path code for CanvasKit and Skwasm
