+++ engine/src/flutter/lib/web_ui/lib/path.dart
@@ -7,8 +7,8 @@ part of ui;
 // For documentation see https://github.com/flutter/flutter/blob/main/engine/src/flutter/lib/ui/painting.dart
 
 abstract class Path {
-  factory Path() => engine.LazyPath(engine.renderer.pathConstructors);
-  factory Path.from(Path source) => engine.LazyPath.fromLazyPath(source as engine.LazyPath);
+  factory Path() => engine.EnginePath(engine.renderer.pathConstructors);
+  factory Path.from(Path source) => engine.EnginePath.fromEnginePath(source as engine.EnginePath);
   PathFillType get fillType;
   set fillType(PathFillType value);
   void moveTo(double x, double y);
@@ -52,7 +52,7 @@ abstract class Path {
   // see https://skia.org/user/api/SkPath_Reference#SkPath_getBounds
   Rect getBounds();
   static Path combine(PathOperation operation, Path path1, Path path2) =>
-      engine.LazyPath.combined(operation, path1 as engine.LazyPath, path2 as engine.LazyPath);
+      engine.EnginePath.combined(operation, path1 as engine.EnginePath, path2 as engine.EnginePath);
 
   PathMetrics computeMetrics({bool forceClosed = false});
 }
