+++ engine/src/flutter/lib/web_ui/lib/src/engine.dart
@@ -18,6 +18,7 @@ library engine;
 export 'engine/alarm_clock.dart';
 export 'engine/app_bootstrap.dart';
 export 'engine/arena.dart';
+export 'engine/backend/path.dart';
 export 'engine/browser_detection.dart';
 export 'engine/canvaskit/canvas.dart';
 export 'engine/canvaskit/canvaskit_api.dart';
@@ -78,7 +79,6 @@ export 'engine/layer/layer_scene_builder.dart';
 export 'engine/layer/layer_tree.dart';
 export 'engine/layer/layer_visitor.dart';
 export 'engine/layer/n_way_canvas.dart';
-export 'engine/lazy_path.dart';
 export 'engine/mouse/context_menu.dart';
 export 'engine/mouse/cursor.dart';
 export 'engine/mouse/prevent_default.dart';
@@ -101,6 +101,8 @@ export 'engine/plugins.dart';
 export 'engine/pointer_binding.dart';
 export 'engine/pointer_binding/event_position_helper.dart';
 export 'engine/pointer_converter.dart';
+export 'engine/primitives/path.dart';
+export 'engine/primitives/path_metrics.dart';
 export 'engine/profiler.dart';
 export 'engine/raw_keyboard.dart';
 export 'engine/renderer.dart';
