+++ engine/src/flutter/lib/web_ui/lib/src/engine/canvaskit/canvas.dart
@@ -45,7 +45,7 @@ class CkCanvas implements LayerCanvas {
   @override
   void clipPath(ui.Path path, {bool doAntiAlias = true}) {
     skCanvas.clipPath(
-      ((path as LazyPath).builtPath as CkPath).skiaObject,
+      ((path as EnginePath).backendPath as CkPath).skiaObject,
       _clipOpIntersect,
       doAntiAlias,
     );
@@ -62,7 +62,7 @@ class CkCanvas implements LayerCanvas {
     final (ui.Path path, ui.Offset offset) = rsuperellipse.toPathOffset();
     translate(offset.dx, offset.dy);
     skCanvas.clipPath(
-      ((path as LazyPath).builtPath as CkPath).skiaObject,
+      ((path as EnginePath).backendPath as CkPath).skiaObject,
       _clipOpIntersect,
       doAntiAlias,
     );
@@ -227,7 +227,7 @@ class CkCanvas implements LayerCanvas {
   @override
   void drawPath(ui.Path path, ui.Paint paint) {
     final SkPaint skPaint = (paint as CkPaint).toSkPaint();
-    skCanvas.drawPath(((path as LazyPath).builtPath as CkPath).skiaObject, skPaint);
+    skCanvas.drawPath(((path as EnginePath).backendPath as CkPath).skiaObject, skPaint);
     skPaint.delete();
   }
 
@@ -269,7 +269,7 @@ class CkCanvas implements LayerCanvas {
     final SkPaint skPaint = (paint as CkPaint).toSkPaint();
     final (ui.Path path, ui.Offset offset) = rsuperellipse.toPathOffset();
     translate(offset.dx, offset.dy);
-    skCanvas.drawPath(((path as LazyPath).builtPath as CkPath).skiaObject, skPaint);
+    skCanvas.drawPath(((path as EnginePath).backendPath as CkPath).skiaObject, skPaint);
     translate(-offset.dx, -offset.dy);
     skPaint.delete();
   }
@@ -286,7 +286,7 @@ class CkCanvas implements LayerCanvas {
   void drawShadow(ui.Path path, ui.Color color, double elevation, bool transparentOccluder) {
     drawSkShadow(
       skCanvas,
-      ((path as LazyPath).builtPath as CkPath).skiaObject,
+      ((path as EnginePath).backendPath as CkPath).skiaObject,
       color,
       elevation,
       transparentOccluder,
